\name{accum.path}
\Rdversion{1.1}
\alias{accum.path}
\title{
function to generate accum.path plot 
}
\description{
work in progress
}
\usage{
accum.path(pems)
}
\arguments{
  \item{pems}{
pems. A pems object or other dataframe
}
}
\details{
work in progress
}
\value{
work in progress
}
\references{
work in progress
}
\author{
karl ropkins
}
\note{
this is a work in progress
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
#temp function
require(lattice)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }

