% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/filter_congestion.R
\name{filter.congestion}
\alias{filter.congestion}
\title{A function to apply congestion data filter on a dataset. The function removes the a day's datat if the running length of the congested data is shorter than 5 min. The congested data is defined as occ>con.occ. The default value of con.occ is 0.3}
\usage{
filter.congestion(tables, con.occ = 0.3)
}
\arguments{
\item{tables}{a list in which each element is a data dataframe for one day's traffic data. It could be the output of read.processed().}

\item{con.occ}{The lower bound of occupancy for congestion. The default value of con.occ is 0.3}
}
\description{
A function to apply congestion data filter on a dataset. The function removes the a day's datat if the running length of the congested data is shorter than 5 min. The congested data is defined as occ>con.occ. The default value of con.occ is 0.3
}
\examples{
tables<-filter.missing(tables, con.occ=0.3)
}

