% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ci.R
\name{get.ci}
\alias{get.ci}
\title{Get candidate intervals given a specific penalty}
\usage{
get.ci(datasets, pen)
}
\arguments{
\item{datasets}{The output from get.total, a list of filtered data where each element correponds to one day's data.}

\item{pen}{Penalty value used in PELT for flow data and occupancy data.}
}
\description{
This function derives characteristics of candidate intervals, including starding and ending point positions, starting and ending time, length, duration, total flow-rate, average occupancy, normalized flow and occupancy gaps.
}
\examples{
ci<-get.CI(datasets,pen1=0.01)
}

