% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.norm.total.R
\name{get.nor.total}
\alias{get.nor.total}
\title{Calculate and convert total flow and occupancy to normalized form.}
\usage{
get.nor.total(datasets)
}
\arguments{
\item{dataset}{The output from read.filtered(), a list of filtered data where each element correponds to one day's data}
}
\description{
Calculate and convert total flow and occupancy to normalized form.
}
\examples{
datasets<-read.filtered(directory="./filtered/",  files="1213891-2011-09.*\\\\.csv")
datasets<-get.total(datasets)
}

