% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ss.R
\name{get.ss}
\alias{get.ss}
\title{A function to identify near-stationary states}
\usage{
get.ss(datasets, mindr, maxdr, hot, gapt)
}
\arguments{
\item{datasets}{The output from get.CI, a list of candidate intervals}

\item{mindr}{Minimum duration of the candidate intervals}

\item{maxdr}{Maximum duration of the candidate intervals}

\item{hot}{The high-occupancy threshold}

\item{gapt}{The gap threshold}
}
\description{
A function to identify near-stationary states
}
\examples{
ss<-get.ss(ci,mindr=4, maxdr=10, hot=0.15, gapt=1)
}

