% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.missing.R
\name{impute.missing}
\alias{impute.missing}
\title{A function to impute missing data using LOWESS based on data index}
\usage{
impute.missing(tables, span = 0.25)
}
\arguments{
\item{tables}{a list in which each element is a data dataframe for one day's traffic data The output of read.processed().}

\item{span}{The parameter in  loess which controls the degress of smoothing}
}
\description{
A function to impute missing data using LOWESS based on data index
}
\examples{
tables<-impute.missing(tables)
}

