% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/parse_pems.R
\name{parse.pems}
\alias{parse.pems}
\title{Parse Pems raw data and pick a subset of VDS}
\usage{
parse.pems(inDir = "./pemsData/", outDir = "./processed/",
  vds = c(1203501, 1203506, 1203534, 1203536, 1212126, 1213891), meta)
}
\arguments{
\item{inDir}{The directory where the pems raw files are located, the raw files are in .gz format}

\item{outDir}{The directory to write the output files, the files are in .csv format}

\item{vds}{A vector of desired VDS's}

\item{meta}{The meta data used to parse the data. It is suggested to read in the meta file by function "read.delim"}
}
\description{
This funtion reads in the pems data and reform it into a "comfortable" format with columns: Time, vds, flow, occ, speed.
}
\examples{
meta <- read.delim("./meta/d12_text_meta_2013_12_14.txt")
parse.pems(meta)
}

