% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/read_processed.R
\name{read.processed}
\alias{read.processed}
\title{Read processed PeMS data}
\usage{
read.processed(directory = "./processed/",
  condition = "1213891__2011_09_.*\\\\.csv")
}
\arguments{
\item{directory}{The directory where the csv files are located, default value is "./processed/"}

\item{condition}{Regex of the file names, default value is "1213891__2011_09_.*\\.csv"}
}
\description{
This funtion is used to read processed PeMS data in csv format (columns: Time, vds, flow, occ, speed) and add date column, convert time to index and remove speed column.
}
\examples{
datasets<-read.processed()
}

