\name{calc.accel}
\Rdversion{1.1}
\alias{calc.accel}
\title{
%%  calculate acceleration 
}
\description{
%%  standard function to calculate acceleration
}
\usage{
calc.accel(pems, speed = velocity, time = local.time, fun.method = "standard.accel", fun.args = NULL, vis.fun = TRUE, fun.output = "pems")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{pems}{
%%     a pems object
}
  \item{speed}{
%%     data series of speed measurements
}
  \item{time}{
%%     data series of timestamps
}
  \item{fun.method}{
%%     calculation method
}
  \item{fun.args}{
%%     additional setting for calculation method
}
  \item{vis.fun}{
%%     visualise the function output
}
  \item{fun.output}{
%%     format of calculation output
}
}
\details{
%%  One of the calc. family of functions
}
\value{
%%  Work in Progress
}
\references{
%% work in progress
}
\author{
%%  karl ropkins
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- This is a test ----

objects()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ calc.accel }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
