% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filter.obs.rate}
\alias{filter.obs.rate}
\title{A function to apply observation rate filter on dataset}
\usage{
filter.obs.rate(tables, obs.rate = 0.95)
}
\arguments{
\item{tables}{a list in which each element is a data dataframe for one day's traffic data, the output of read.processed().}

\item{obs.rate}{The threshold of minimum daily obervation rate}
}
\description{
A function to apply observation rate filter on dataset
}
\examples{
tables<-filter.obs.rate(tables, obs.rate=0.95)
}

