% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.ss}
\alias{get.ss}
\title{get steady state time sequence}
\usage{
get.ss(datasets, pen1 = 2, pen2 = 2, len = 30)
}
\arguments{
\item{datasets}{The output from get.total, a list of filtered data where each element correponds to one day's data}

\item{pen1}{Penalty value used in PELT for flow data}

\item{pen2}{Penalty value used in PELT for occupancy data}

\item{len}{The minimum duration of the steady state}
}
\description{
get steady state time sequence
}
\examples{
ss<-get.ss(datasets,pen1=2,pen2=2)
}

