% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.total}
\alias{get.total}
\title{Calulate the total flow and total occupancy accross lanes}
\usage{
get.total(datasets)
}
\arguments{
\item{dataset}{The output from read.filtered, a list of filtered data where each element correponds to one day's data}
}
\description{
Calulate the total flow and total occupancy accross lanes
}
\examples{
datasets<-read.filtered(directory="./filtered/",  files="1213891-2011-09.*\\\\.csv")
datasets<-get.total(datasets)
}

