% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.filtered.R
\name{read.filtered}
\alias{read.filtered}
\title{Read filtered PeMS data}
\usage{
read.filtered(directory = "./filtered/",
  files = "1213891-2011-09.*\\\\.csv")
}
\arguments{
\item{directory}{The directory where the csv files are located, default value is "./filtered/"}

\item{condition}{Regex of the file names, default value is "1213891-2011-09.*\\.csv"}
}
\description{
This funtion is used to read filtered PeMS data in csv format with column: Data, Index, vd, flow, occ.
}
\examples{
datasets<-read.filtered(directory="./filtered/",  files="1213891-2011-09.*\\\\.csv")
}

