% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.filtered}
\alias{write.filtered}
\title{A function write the filtered data into csv file}
\usage{
write.filtered(tables)
}
\arguments{
\item{tables}{a list in which each element is a data dataframe for one day's traffic data, the output of read.processed().}
}
\description{
A function write the filtered data into csv file
}
\examples{
write.filtered(tables)
}

